requirements.txt
openai==1.86.0
python-dotenv==1.1.0
streamlit==1.32.0
